+++ app/lib/src/stores/password.dart
@@ -1,0 +1,2 @@
+import 'package:mobx/mobx.dart';
+import 'package:generated_io/generated_io.dart';
