+++ app/lib/src/views/loading/loading.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/widgets.dart';
 import 'package:stashall/src/routes/passwords_route.dart';
 import 'package:stashall/src/routes/signin_route.dart';
