+++ app/pubspec.yaml
@@ -11,6 +11,8 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^0.1.2
+  mobx: ^0.2.0
+  flutter_mobx: ^0.2.0+1
   rc_router:
     git: https://github.com/rushio-consulting/rc_router.git
   generated_io:
